Reorganize classes by functionality
