+++ src/main/java/com/google/idea/perf/methodtracer/Autocomplete.kt
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
+import com.google.idea.perf.fuzzyMatch
+import com.google.idea.perf.fuzzySearch
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.CompletionUtilCore
