+++ src/main/java/com/google/idea/perf/methodtracer/CallTree.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 /** A call tree, represented recursively. */
 interface CallTree {
