+++ src/main/java/com/google/idea/perf/methodtracer/CallTreeManager.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 // Things to improve:
 // - GC the state for dead threads.
