+++ src/main/java/com/google/idea/perf/methodtracer/Tracepoint.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 import java.util.concurrent.atomic.AtomicInteger
