+++ src/main/java/com/google/idea/perf/methodtracer/TracerMethodTransformer.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 import com.google.idea.perf.agent.Trampoline
 import com.intellij.openapi.diagnostic.Logger
