+++ src/main/java/com/google/idea/perf/methodtracer/TracerView.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -71,7 +71,7 @@ class TracerDialog : DialogWrapper(null, null, false, IdeModalityType.IDE, false
 
     override fun createCenterPanel(): JComponent = TracerView(disposable)
     override fun createContentPaneBorder(): Border? = null // No border.
-    override fun getDimensionServiceKey(): String = "com.google.idea.perf.Tracer"
+    override fun getDimensionServiceKey(): String = "com.google.idea.perf.methodtracer.Tracer"
     override fun createActions(): Array<Action> = emptyArray()
 }
