+++ src/test/java/com/google/idea/perf/methodtracer/CallTreeTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
