+++ src/test/java/com/google/idea/perf/methodtracer/TracerCommandTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.idea.perf
+package com.google.idea.perf.methodtracer
 
 import org.junit.Test
 import kotlin.test.assertEquals
